GuideWire Update
1. TIF Class -> OCTImaging Class
2. Core쪽에서 Detection 실패 시 더미 값 추가 (Frameinfo <-> FrameNum 매치 되어야 함)
3. GuideWire Center Detection Logic 수정 필요, 현재 Angle을 15로 고정하였음

+++ RayCore/RayCore/ImagingSession.h
@@ -49,7 +49,7 @@ private:
 	std::vector<std::vector<cv::Mat>> m_vSidebranch;
 	std::vector<cv::Mat> m_vStent;
 	std::vector<cv::Mat> m_vGuidewire;
-	std::vector<std::vector<double>> m_vGuidewireRadius;
+	std::vector<std::vector<float>> m_vGuidewireRadius;
 	char* m_pVolumeData;
 
 	int m_zOffset;
